Added more reference materials  aka the javascript from access plus It seems like someone was working on an export to csv but it never really got off of the ground.
Refactored the the Course class so that it now holds an additional subclass MeetingTime(). This was done because the logic for keeping track of standard classes/Arranged Classes/Classes with multiple meeting times was getting kind of crazy.
In light of this I was able to remove the lambda expressions that I was using  in build_date time(). It's now a boring method like the rest.

 Continued incremental improvements to the IcsGenerator() class, it now can add classes with multiple meeting times and should properly handle Arranged classes aswell

Added additional TODOs as per usual

This commit is broken but I need to switch dev machines so here it is prematurely